# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TAOX API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/products/": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get All Products",
                "description": "Lista todos los productos (requiere autenticación)",
                "operationId": "get_all_products_products__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductDTO"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Products Products  Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create Product",
                "description": "Crea un nuevo producto (requiere rol de administrador)",
                "operationId": "create_product_products__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
        "/products/{product_id}": {
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update Product",
                "description": "Actualiza un producto (requiere rol de administrador)",
                "operationId": "update_product_products__product_id__put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Product Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete Product",
                "description": "Elimina un producto (requiere rol de administrador)",
                "operationId": "delete_product_products__product_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Product Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/movements/": {
            "get": {
                "tags": [
                    "movements"
                ],
                "summary": "Get All Movements",
                "description": "Obtiene todos los movimientos de una empresa",
                "operationId": "get_all_movements_movements__get",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductMovementResponseDTO"
                                    },
                                    "title": "Response Get All Movements Movements  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "movements"
                ],
                "summary": "Create Movement",
                "description": "Crea un nuevo movimiento. Automáticamente genera alertas si hay cambios de stock crítico",
                "operationId": "create_movement_movements__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductMovementCreateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductMovementResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movements/{id_movimiento}": {
            "get": {
                "tags": [
                    "movements"
                ],
                "summary": "Get Movement By Id",
                "description": "Obtiene un movimiento por ID",
                "operationId": "get_movement_by_id_movements__id_movimiento__get",
                "parameters": [
                    {
                        "name": "id_movimiento",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Movimiento"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductMovementResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "movements"
                ],
                "summary": "Update Movement",
                "description": "Actualiza un movimiento",
                "operationId": "update_movement_movements__id_movimiento__put",
                "parameters": [
                    {
                        "name": "id_movimiento",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Movimiento"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductMovementUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductMovementResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "movements"
                ],
                "summary": "Delete Movement",
                "description": "Elimina un movimiento",
                "operationId": "delete_movement_movements__id_movimiento__delete",
                "parameters": [
                    {
                        "name": "id_movimiento",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Movimiento"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Company Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movements/{id_movimiento}/status": {
            "patch": {
                "tags": [
                    "movements"
                ],
                "summary": "Update Movement Status",
                "description": "Actualiza el estado de un movimiento",
                "operationId": "update_movement_status_movements__id_movimiento__status_patch",
                "parameters": [
                    {
                        "name": "id_movimiento",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id Movimiento"
                        }
                    },
                    {
                        "name": "nuevo_estado",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Nuevo Estado"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductMovementResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateProductDTO": {
                "properties": {
                    "id_product": {
                        "type": "string",
                        "title": "Id Product"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "generic_name": {
                        "type": "string",
                        "title": "Generic Name"
                    },
                    "price": {
                        "type": "number",
                        "title": "Price"
                    },
                    "unit_measure": {
                        "type": "string",
                        "title": "Unit Measure"
                    },
                    "unit_price": {
                        "type": "number",
                        "title": "Unit Price"
                    },
                    "min_unit_price": {
                        "type": "number",
                        "title": "Min Unit Price"
                    },
                    "lead_time_days": {
                        "type": "integer",
                        "title": "Lead Time Days"
                    },
                    "restorage": {
                        "type": "string",
                        "title": "Restorage"
                    },
                    "company_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Company Id"
                    }
                },
                "type": "object",
                "required": [
                    "id_product",
                    "name",
                    "generic_name",
                    "price",
                    "unit_measure",
                    "unit_price",
                    "min_unit_price",
                    "lead_time_days",
                    "restorage",
                    "company_id"
                ],
                "title": "CreateProductDTO"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProductDTO": {
                "properties": {
                    "id_product": {
                        "type": "string",
                        "title": "Id Product"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "generic_name": {
                        "type": "string",
                        "title": "Generic Name"
                    },
                    "price": {
                        "type": "number",
                        "title": "Price"
                    },
                    "unit_measure": {
                        "type": "string",
                        "title": "Unit Measure"
                    },
                    "unit_price": {
                        "type": "number",
                        "title": "Unit Price"
                    },
                    "min_unit_price": {
                        "type": "number",
                        "title": "Min Unit Price"
                    },
                    "lead_time_days": {
                        "type": "integer",
                        "title": "Lead Time Days"
                    },
                    "restorage": {
                        "type": "string",
                        "title": "Restorage"
                    },
                    "company_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Company Id"
                    }
                },
                "type": "object",
                "required": [
                    "id_product",
                    "name",
                    "generic_name",
                    "price",
                    "unit_measure",
                    "unit_price",
                    "min_unit_price",
                    "lead_time_days",
                    "restorage",
                    "company_id"
                ],
                "title": "ProductDTO"
            },
            "ProductMovementCreateDTO": {
                "properties": {
                    "codigo_producto": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Codigo Producto",
                        "description": "Código del producto"
                    },
                    "id_proceso_origen": {
                        "type": "integer",
                        "title": "Id Proceso Origen",
                        "description": "ID del proceso de origen"
                    },
                    "id_proceso_destino": {
                        "type": "integer",
                        "title": "Id Proceso Destino",
                        "description": "ID del proceso de destino"
                    },
                    "cantidad": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Cantidad",
                        "description": "Cantidad a mover"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas",
                        "description": "Notas adicionales del movimiento"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa",
                        "description": "ID de la empresa propietaria"
                    }
                },
                "type": "object",
                "required": [
                    "codigo_producto",
                    "id_proceso_origen",
                    "id_proceso_destino",
                    "cantidad",
                    "id_empresa"
                ],
                "title": "ProductMovementCreateDTO",
                "description": "DTO para crear un nuevo movimiento",
                "example": {
                    "cantidad": 10.5,
                    "codigo_producto": "CHEM001",
                    "id_empresa": "company-123",
                    "id_proceso_destino": 2,
                    "id_proceso_origen": 1,
                    "notas": "Movimiento de producción a préstamo"
                }
            },
            "ProductMovementResponseDTO": {
                "properties": {
                    "id_movimiento": {
                        "type": "integer",
                        "title": "Id Movimiento"
                    },
                    "codigo_producto": {
                        "type": "string",
                        "title": "Codigo Producto"
                    },
                    "id_proceso_origen": {
                        "type": "integer",
                        "title": "Id Proceso Origen"
                    },
                    "id_proceso_destino": {
                        "type": "integer",
                        "title": "Id Proceso Destino"
                    },
                    "cantidad": {
                        "type": "number",
                        "title": "Cantidad"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    },
                    "id_empresa": {
                        "type": "string",
                        "title": "Id Empresa"
                    },
                    "estado": {
                        "type": "string",
                        "title": "Estado"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id_movimiento",
                    "codigo_producto",
                    "id_proceso_origen",
                    "id_proceso_destino",
                    "cantidad",
                    "notas",
                    "id_empresa",
                    "estado",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProductMovementResponseDTO",
                "description": "DTO para respuestas de movimiento",
                "example": {
                    "cantidad": 10.5,
                    "codigo_producto": "CHEM001",
                    "created_at": "2025-03-01T10:00:00",
                    "estado": "completado",
                    "id_empresa": "company-123",
                    "id_movimiento": 1,
                    "id_proceso_destino": 2,
                    "id_proceso_origen": 1,
                    "notas": "Movimiento de producción a préstamo",
                    "updated_at": "2025-03-01T10:30:00"
                }
            },
...
            "UpdateProductDTO": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "generic_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Generic Name"
                    },
                    "price": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price"
                    },
                    "unit_measure": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unit Measure"
                    },
                    "unit_price": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unit Price"
                    },
                    "min_unit_price": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Min Unit Price"
                    },
                    "lead_time_days": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Lead Time Days"
                    },
                    "restorage": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Restorage"
                    }
                },
                "type": "object",
                "title": "UpdateProductDTO"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}